Added support for command callbacks.
Added a new addCommand(Object object, Object callable) interface that accepts one of JCallable<T>, Callable or Runnable, as a callback to be executed after the given command is successfully parsed.

In order to propagate exceptions from the callbacks more cleanly, added a parseChecked() method in addition to the standard parse(), with a throws declaration. Any exceptions thrown by the callback would be propagated as-is the the caller of parseChecked().

Extended the documentation to include samples for using command names and aliases in annotations, and using callables.
Note: documentation makes a hard-coded reference to ver 1.18.