Removed CommaSeparated converters and added -list option to Parameter.
Previous commit I added some CommaSeparated converters which
allowed for building of java.util.List objects from comma-
separated values.  It was an ugly implementation as for
each new converter we had to add a second converter if
wanting to use comma-separated lists of that type.

I've removed those CommaSeparated converters and added a -list
option to parameter which defaults to false.  If we
specify list=true in the parameter annotation we build
an ArrayList utilising the converter specified.

For example:
@Parameter(..... converter = BigDecimalConverter.class, list = true)
will generate a List<BigDecimal> by parsing each element from the
user input "123.45, 67.89" as BigDecimal objects.

Commas can be backslash-escaped if required within an individual
element.