Keep the parent's JCommander bundle when creating child JCommander objects for commands
This commit should fix a bug in which commands added to the JCommander object do not have a resource bundle associated with them and thus appear to not have descriptions if their descriptions are provided using a bundle key and not directly in the annotation. Unfortunately, the parameter descriptions are created in the constructor and later change in the bundle does not affect them.