No ParameterException if a default value assigned for required params
If a default value is assigned to a required parameter on its
initialization, JCommander won't throw ParameterException even if no
value is given by user.

Due to this fix JCommanderTest#requiredFields1Fail did not fail anymore
since Args1#verbose was a required parameter with a default value. To
distinguish the two cases when a required parameter has no default value
and has a default value, this commit removes the default value from
Args1#verbose.