Allow user to retrieve unknown command
When an unknown command is parsed, the MissingCommandException is
raised. However, there is no way for the library user to access the
erroneous command (except parsing either the arguments or the error
message).
Such access is helpfull, for instance to provide a better error message,
hinting to close commands as git does (try "git pulll").