Fix error message when failing to convert a Path
When a parameter is of type Path, the PathConverter is used.
If the path is not a valid Path, for instance when the user input is an
uri, a stack trace is shown, where a proper error message is expected.

When converting a path, catch the InvalidPathException that can be
thrown, and format the message as it is done in other converters.