Use return type of method to determine parameter type
Shouldn't the return type of the method be used to determine the parameter type? I ran into this issue when experimenting with enums (generic enums to be precise...) and when I tried to annotate methods with `@Parameter` instead of class fields. Using the first parameter of a method call does not make sense to me unless I am missing something. The annotation should be on getters anyways without any method parameters.

Parameters on methods instead of class fields does not seem to be something that is documented, but by the `@Parameter` annotation with `@Target({ FIELD, METHOD })` and from various bug reports and PR's I learned that it is possible.